Improve security & code quality
