+++ app/controllers/carts_controller.rb
@@ -5,7 +5,7 @@ class CartsController < ApplicationController
   def destroy
     @cart.destroy 
     session[:cart_id] = nil
-    redirect_to(home_url, :notice => 'Your cart is currently empty') 
+    redirect_to(home_url, notice: 'Your cart is currently empty') 
   end
 
 end
