--- app/controllers/concerns/set_cart.rb
@@ -1,12 +1,0 @@
-module SetCart
-
-  private
-
-    def set_cart
-      @cart = Cart.find(session[:cart_id])
-    rescue ActiveRecord::RecordNotFound
-      @cart = Cart.create
-      session[:cart_id] = @cart.id
-    end
-
-end
