+++ app/models/cart.rb
@@ -1,5 +1,5 @@
 class Cart < ApplicationRecord
-  has_many :order_items, :dependent => :destroy
+  has_many :order_items, dependent: :destroy
 
   def add_product(product_id)
     current_item = order_items.find_by_product_id(product_id) 
@@ -12,7 +12,7 @@ class Cart < ApplicationRecord
   end
 
   def total_price
-    order_items.to_a.sum { |item| item.total_price }
+    order_items.to_a.sum(&:total_price)
   end
   
 end
