+++ app/views/carts/_cart.html.haml
@@ -6,11 +6,11 @@
         %td.col-sm-2= "#{item.quantity}x"
         %td.col-sm-4= item.product.name
         %td.col-sm-2.item_price= number_to_currency(item.total_price)
-        %td.col-sm-4= button_to 'clear', order_item_path(item.id), class: "clear-button btn btn-warning", :method => :delete, :confirm => 'Are you sure?'
+        %td.col-sm-4= button_to 'clear', order_item_path(item.id), class: "clear-button btn btn-warning", method: :delete, confirm: 'Are you sure?'
     %tr.row
       %td.col-sm-3.offset-sm-3 Total:
       %td.col-sm-1= number_to_currency(@cart.total_price)
   .row.col-md-12.justify-content-center
-    = button_to 'Checkout!', new_order_path, class:"btn btn-success", :method => :get
+    = button_to 'Checkout!', new_order_path, class:"btn btn-success", method: :get
   .row.col-md-12.justify-content-center
-    = button_to 'Empty cart', @cart, class:"btn btn-primary", :method => :delete, :confirm => 'Are you sure?'
+    = button_to 'Empty cart', @cart, class:"btn btn-primary", method: :delete, confirm: 'Are you sure?'
