+++ app/views/layouts/_sorting_bar.haml
@@ -24,7 +24,7 @@
     .score
       .slidecontainer
         .range-slider
-          %input{ :max => '100', :min => '1', :step => '1', :type => 'range', :value => '1' }/
-          %input{ :max => '100', :min => '1', :step => '1', :type => 'range', :value => '100' }/
+          %input{ max: '100', min: '1', step: '1', type: 'range', value: '1' }/
+          %input{ max: '100', min: '1', step: '1', type: 'range', value: '100' }/
     #slider-value-2.score
       $100
