+++ app/views/layouts/top_header/navbar/_logged_in_panel.haml
@@ -1,7 +1,7 @@
 %li.nav-item.active
   %span.middot.nav-link ·
 %li.nav-item.dropdown
-  %a.dropdown-toggle.home-page-text.nav-link{ 'data-toggle' => 'dropdown', :href => '#' }
+  %a.dropdown-toggle.home-page-text.nav-link{ 'data-toggle' => 'dropdown', href: '#' }
     Account
   %ul.dropdown-menu.home-page-text
     %li= link_to 'Profile', current_user
