+++ app/views/layouts/top_header/navbar/_registration.html.haml
@@ -1,8 +1,8 @@
 %li.nav-item.active
   %span.middot.nav-link ·
 %li.nav-item
-  = link_to 'Log in', login_path, :class =>'nav-link home-page-text'
+  = link_to 'Log in', login_path, class: 'nav-link home-page-text'
 %li.nav-item.active
   %span.middot.nav-link ·
 %li.nav-item
-  = link_to 'Sign up', new_user_path, :class =>'nav-link home-page-text'
+  = link_to 'Sign up', new_user_path, class: 'nav-link home-page-text'
