+++ app/views/users/show.html.haml
@@ -2,7 +2,6 @@
   %aside.col-md-4
     %section.user_info
       %h1= @user.name
-
   .col-md-4.w-50
     %h2 Your orders!
     - @user.orders.each do |order|
@@ -22,6 +21,3 @@
               %td
               %td.col-md-2= "Total: #{order.total_price}"
               %td.col-md-2= "Status: #{order.status.capitalize}"
-
-                
-
